pyproject.toml
[project]
name = "cabamodel"
version = "0.1.0"
description = "High-performance Modular Gemini-native Agent Architecture using Google ADK."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.23.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "tenacity>=9.0.0",
]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
